backend/flask/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
numpy
opencv-python-headless
scikit-image
pillow